restructure tags folder and imports, split tags manager out from zetkin section. create new tag manager with section
